Move routes/course to database/model/course
