+++ server/src/app.module.ts
@@ -9,7 +9,7 @@ import { UserSeederModule } from './database/seeders/user/user-seeder.module';
 import { EventEmitterModule } from '@nestjs/event-emitter';
 import { EventModule } from './events/event.module';
 import { AccountModule } from './routes/account/account.module';
-import { CoursesModule } from './routes/course/courses.module';
+import { CoursesModule } from './database/models/course/course.module';
 import { CourseSeederModule } from './database/seeders/course/course-seeder.module';
 
 @Module({
