+++ server/src/database/models/course/course.module.ts
@@ -1,7 +1,7 @@
 import { Module } from "@nestjs/common";
 import { MongooseModule } from "@nestjs/mongoose/dist/mongoose.module";
-import { CourseSchema } from '../../database/schemas/course.schema';
-import { CoursesService } from './courses.service';
+import { CourseSchema } from '../../schemas/course.schema';
+import { CoursesService } from './course.repository';
 
 @Module({
     imports: [MongooseModule.forFeature([{ name: 'Course', schema: CourseSchema }])],
