+++ server/src/database/models/course/dto/update.dto.ts
@@ -1,6 +1,6 @@
-import CourseLevel from '../../../common/constants/course-level';
-import CourseRating from '../../../common/constants/course-rating';
-import CourseTopic from '../../../common/constants/course-topic';
+import CourseLevel from '../../../../common/constants/course-level';
+import CourseRating from '../../../../common/constants/course-rating';
+import CourseTopic from '../../../../common/constants/course-topic';
 
 export class UpdateCourseDTO implements Readonly<UpdateCourseDTO> {
     authorName?: string;
