+++ server/src/database/models/course/interfaces/ICourse.ts
@@ -1,7 +1,7 @@
 import { Document } from 'mongoose';
-import CourseLevel from '../../../common/constants/course-level';
-import CourseRating from '../../../common/constants/course-rating';
-import CourseTopic from '../../../common/constants/course-topic';
+import CourseLevel from '../../../../common/constants/course-level';
+import CourseRating from '../../../../common/constants/course-rating';
+import CourseTopic from '../../../../common/constants/course-topic';
 
 export interface ICourse extends Document {
     readonly id: string;
