+++ server/src/database/seeders/course/course-seeder.module.ts
@@ -1,6 +1,6 @@
 import { Module } from "@nestjs/common";
 import { CommandModule } from "nestjs-command";
-import { CoursesModule } from '../../../routes/course/courses.module';
+import { CoursesModule } from '../../models/course/course.module';
 import { UserModule } from '../../models/user/user.module';
 import { CourseSeeder } from './course.seeder';
