+++ server/src/database/seeders/course/course.seeder.ts
@@ -2,8 +2,8 @@ import { Command, Option } from 'nestjs-command';
 import { Injectable } from '@nestjs/common';
 import * as faker from 'faker';
 import { logger } from '../../../common/utils/logger';
-import { CreateCourseDTO } from '../../../routes/course/dto/create.dto';
-import { CoursesService } from '../../../routes/course/courses.service';
+import { CreateCourseDTO } from '../../models/course/dto/create.dto';
+import { CoursesService } from '../../models/course/course.repository';
 import { Constants } from '../../../common/constants';
 import { IUser } from '../../models/user/interfaces/IUser';
 import { UserRepository } from '../../models/user/user.repository';
